chore: Restructure `source_data`
