+++ fetch_data/source-data/sources/imerg.py
@@ -15,8 +15,9 @@ import shutil
 from datetime import date
 
 import earthaccess
-import models
-from configs.settings import Settings, set_logging
+
+from . import models
+from .configs.settings import Settings, set_logging
 
 set_logging()
 logger = logging.getLogger(__name__)
